requirements.txt
asgiref==3.7.2
blinker==1.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pillow==10.2.0
psycopg==3.1.16
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-decouple==3.8
pytz==2023.3.post1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
